Add AppSync and Cognito integrations
This adds support for the AppSync endpoint I provisioned
(a big TODO is still to get all the infra codified in
terraform, but... eh... another day). There's a little
bit of jank around this currently (reason for the
`(... as unknown) as ApolloLink)` bullshit), but It Works™
so calling it a day there.

I also added the relevant Amplify bits to support user sign
in/up. Not sure how I want to handle the platform + gamertag
bit long-term since I'm being lazy in how I handle user profiles
(i.e., they don't exist and I'm just cramming a couple custom
attribs onto the Cognito user objects). Dropdowns in Amplify
sign-up flows aren't really supported, but there _is_ a pattern
for that by cribbing some of the country code stuff they implemented
for phone numbers in the sign-up flow. So it's possible, just
annoying. Another option is to let users modify this after
signing up, and this looks like it'll be pretty straightforward with
`Auth.updateUserAttributes(...)`. So that's promising. Probably
also the objectively better option since folks can change platforms
over time. Doesn't really solve for the "I have an XBox and a PC"
use case, but I don't think I care.